+++ System/delay/delay.h
@@ -2,7 +2,7 @@
 #define _DELAY_H
 #include <sys.h>
 
-void delay_ms(u16 nms);
+void delay_ms(u32 nms);
 void delay_us(u32 nus);
 #endif
